requirements.txt
pandas~=1.5.3
openai~=0.27.0
matplotlib~=3.7.1
tqdm~=4.64.1
py-rouge~=1.1
bert-score~=0.3.13
nltk~=3.8.1
scipy~=1.10.1
backoff